requirements.txt
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.3.0
exceptiongroup==1.2.0
fastapi==0.103.2
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httptools==0.6.0
idna==3.6
importlib-metadata==6.7.0
loguru==0.7.2
motor==3.3.2
passlib==1.7.4
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.1
urllib3==2.0.7
uvicorn==0.22.0
watchfiles==0.20.0
websockets==11.0.3
win32-setctime==1.1.0
zipp==3.15.0
